+++ packages/example-antd-management-fast-framework/src/businessComponents/Drawers/SimpleVerticalFlexDrawer/codeSource.js
@@ -1,28 +1,25 @@
 export const code = `import { Checkbox } from 'antd';
 
 import { connect } from 'easy-soft-dva';
-import { convertCollection, getValueByKey } from 'easy-soft-utility';
 
-import { cardConfig, drawerConfig } from 'antd-management-fast-common';
-import { iconBuilder, SyntaxHighlighter } from 'antd-management-fast-component';
+import { drawerConfig } from 'antd-management-fast-common';
+import { SyntaxHighlighter } from 'antd-management-fast-component';
 import {
   DataDrawer,
   switchControlAssist,
 } from 'antd-management-fast-framework';
 
-import { fieldData } from '../../../businessData/data';
-
 import { code } from './codeSource';
 
-const { BaseUpdateDrawer } = DataDrawer;
+const { BaseVerticalFlexDrawer } = DataDrawer;
 
-const visibleFlag = '7476eba9e3bf442ab7655e7b41c40360';
+const visibleFlag = '7fc0c679771c42cb9d8855fa00bb2104';
 
 @connect(({ simple, schedulingControl }) => ({
   simple,
   schedulingControl,
 }))
-class SimpleEditModel extends BaseUpdateDrawer {
+class SimpleVerticalFlexDrawer extends BaseVerticalFlexDrawer {
   // showCallProcess = true;
 
   static open() {
@@ -63,135 +60,6 @@ class SimpleEditModel extends BaseUpdateDrawer {
     return '编辑信息';
   };
 
-  fillInitialValuesAfterLoad = ({
-    metaData = null,
-    // eslint-disable-next-line no-unused-vars
-    metaListData = [],
-    // eslint-disable-next-line no-unused-vars
-    metaExtra = null,
-    // eslint-disable-next-line no-unused-vars
-    metaOriginalData = null,
-  }) => {
-    const values = {};
-
-    if (metaData != null) {
-      values[fieldData.title.name] = getValueByKey({
-        data: metaData,
-        key: fieldData.title.name,
-      });
-
-      values[fieldData.description.name] = getValueByKey({
-        data: metaData,
-        key: fieldData.description.name,
-      });
-    }
-
-    return values;
-  };
-
-  establishCardCollectionConfig = () => {
-    const { metaData } = this.state;
-
-    return {
-      list: [
-        {
-          title: {
-            icon: iconBuilder.contacts(),
-            text: '基本信息',
-          },
-          items: [
-            {
-              lg: 24,
-              type: cardConfig.contentItemType.input,
-              fieldData: fieldData.title,
-              require: true,
-            },
-          ],
-          instruction: {
-            title: '局部操作说明',
-            showDivider: false,
-            showNumber: true,
-            list: [
-              {
-                text: '这是一些操作说明1',
-              },
-              {
-                text: '这是一些操作说明2',
-              },
-            ],
-          },
-        },
-        {
-          title: {
-            icon: iconBuilder.contacts(),
-            text: '简介描述',
-          },
-          items: [
-            {
-              lg: 24,
-              type: cardConfig.contentItemType.textarea,
-              fieldData: fieldData.description,
-            },
-          ],
-          instruction: [
-            {
-              title: '局部操作说明1',
-              showDivider: false,
-              showNumber: true,
-              list: [
-                {
-                  text: '这是一些操作说明1',
-                },
-                {
-                  text: '这是一些操作说明2',
-                },
-              ],
-            },
-            {
-              title: '局部操作说明2',
-              showDivider: false,
-              showNumber: true,
-              list: [
-                {
-                  text: '这是一些操作说明1',
-                },
-                {
-                  text: '这是一些操作说明2',
-                },
-              ],
-            },
-          ],
-        },
-        {
-          title: {
-            icon: iconBuilder.contacts(),
-            text: '其他信息',
-          },
-          items: [
-            {
-              type: cardConfig.contentItemType.onlyShowInputDatetime,
-              fieldData: fieldData.createTime,
-              value: getValueByKey({
-                data: metaData,
-                key: fieldData.createTime.name,
-                convert: convertCollection.datetime,
-              }),
-            },
-            {
-              type: cardConfig.contentItemType.nowTime,
-              fieldData: fieldData.updateTime,
-              value: getValueByKey({
-                data: metaData,
-                key: fieldData.updateTime.name,
-                convert: convertCollection.datetime,
-              }),
-            },
-          ],
-        },
-      ],
-    };
-  };
-
   establishHelpConfig = () => {
     return {
       title: '操作提示',
@@ -204,6 +72,10 @@ class SimpleEditModel extends BaseUpdateDrawer {
         },
       ],
     };
+  };
+
+  renderPresetContentContainorInnerTop = () => {
+    return 1212;
   };
 
   renderOverlayContent = () => {
@@ -222,5 +94,5 @@ class SimpleEditModel extends BaseUpdateDrawer {
   };
 }
 
-export default SimpleEditModel;
+export { SimpleVerticalFlexDrawer };
 `;